+++ libft.h
@@ -90,11 +90,11 @@ int				ft_tolower(int c);
 
 int				ft_atoi(const char *str);
 char			*ft_itoa(int n);
-char			*ft_itoa_base(long long nbr, int base);
+char			*ft_itoa_base(intmax_t nbr, int base);
 int				ft_sqrt(int nb);
 long long		ft_power(int n, int p);
 int				ft_fibonacci(int nb);
-int				ft_nbrlen(int nbr, int base);
+int				ft_nbrlen(intmax_t nbr, int base);
 int				ft_abs(int n);
 
 /*
